Collapse four URL canonicalization copies into crate::urlnorm
playlist.rs, scan.rs, and history.rs (twice) each hand-rolled the same
strip-fragment/strip-default-port logic feeding dedup keys, history
identity keys, and shared-URL scan caching. A future edit to one copy
would silently desynchronize history diffs from scan dedup. One module
now owns the three shapes (parsed-URL identity, raw stream URL with
fallback, http/https-only identity).
